cleanup methods and files
